+++ src/apps/toolbar/modules/Workspaces/index.tsx
@@ -116,7 +116,6 @@ export function WorkspacesModule({ module }: Props) {
   return (
     <AnimatedDropdown
       animationDescription={{
-        maxAnimationTimeMs: 500,
         openAnimationName: 'ft-bar-item-context-menu-open',
         closeAnimationName: 'ft-bar-item-context-menu-close',
       }}