+++ src/apps/toolbar/modules/shared/store/app.ts
@@ -5,7 +5,7 @@ import {
   DesktopFolder,
   DocumentsFolder,
   DownloadsFolder,
-  invoke,
+  LanguageList,
   MusicFolder,
   PicturesFolder,
   RecentFolder,
@@ -16,8 +16,9 @@ import {
   VideosFolder,
 } from '@seelen-ui/lib';
 import { Placeholder, ToolbarItem } from '@seelen-ui/lib/types';
+import { invoke } from '@tauri-apps/api/core';
 
-import { PowerPlan, RootState } from './domain';
+import { Battery, PowerPlan, PowerStatus, RootState } from './domain';
 
 import { StateBuilder } from '../../../../shared/StateBuilder';
 
@@ -50,7 +51,7 @@ const initialState: RootState = {
     batteryLifeTime: -1,
     batteryFullLifeTime: -1,
   },
-  powerPlan: PowerPlan.Balanced,
+  powerPlan: PowerPlan.Unknown,
   batteries: [],
   workspaces: [],
   activeWorkspace: null,
@@ -64,6 +65,7 @@ const initialState: RootState = {
   mediaInputs: [],
   notifications: [],
   colors: UIColors.default().inner,
+  languages: [],
 };
 
 export const RootSlice = createSlice({
@@ -125,6 +127,16 @@ export const Selectors = StateBuilder.compositeSelector(initialState);
 
 // no core things that can be lazy loaded to improve performance
 export async function lazySlice(d: Dispatch) {
+  invoke<PowerStatus>(SeelenCommand.GetPowerStatus).then((status) =>
+    d(RootActions.setPowerStatus(status)),
+  );
+  invoke<PowerPlan>(SeelenCommand.GetPowerMode).then((plan) => d(RootActions.setPowerPlan(plan)));
+  invoke<Battery[]>(SeelenCommand.GetBatteries).then((batteries) =>
+    d(RootActions.setBatteries(batteries)),
+  );
+
+  LanguageList.getAsync().then((list) => d(RootActions.setLanguages(list.asArray())));
+
   const obj = {
     userRecentFolder: (await RecentFolder.getAsync()).asArray(),
     userDesktopFolder: (await DesktopFolder.getAsync()).asArray(),